interpret: refactor function call handling to be better-abstracted
